requirements.txt
Django==1.11.6
Pillow==4.3.0
psycopg2==2.7.3
gunicorn==19.7.1
django-embed-video==1.1.2
django-tinymce==2.6.0